requirements.txt
aiohttp[speedups]
databases[asyncmy]
elasticsearch
fastapi
orjson
osu.py
python-dateutil
starlette_exporter
uvicorn[standard]
